### 2022-11-27 #4975 @editorialbot
[PRE REVIEW]: Multi-view-AE: A Python package for multi-view autoencoder models
**Submitting author:** @alawryaguila (<a href="http://orcid.org/0000-0003-0727-3274">Ana Lawry Aguila</a>)
**Repository:** https://github.com/alawryaguila/multi-view-AE
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
