### 2022-11-28 dcfcfe1
small paper fixes
