+++ docs/JOSS/paper.md
@@ -32,9 +32,9 @@ bibliography: paper.bib
 ---
 Often, data can be naturally described via multiple views or modalities. For example, we could consider an image and the corresponding text as different modalities. These modalities contain complementary information which can be modelled jointly using multi-view methods. The joint modelling of multiple modalities has been explored in many research fields such as medical imaging [@Serra2018], chemistry [@Sjostrom1983], and natural language processing [@Sadr2020]. 
 
-Autoencoders are unsupervised generative models which learn low dimensional latent representations of complex data. The autoencoder framework consists of two mappings; the encoder which embeds information from the input space into a latent space, and a decoder which transforms point estimates from the latent space back into in the input space. Autoencoders have been successful in down-stream tasks such as classification [@Creswell2017], outlier detection [@An2015], and data generation [@Wei2020]. There exist many frameworks for extending autoencoders to multiple modalities. By far the most popular group of multi-view autoencoder frameworks are multi-view extensions of Variational Autoencoders (VAEs). However, there are also other multi-view autoencoder frameworks, such as multi-view Adversarial Autoencoders (AAEs) [@Wang2019]. Even within multi-view VAEs there are vast modelling differences to be considered when choosing the best model for the task at hand. 
+Autoencoders are unsupervised generative models which learn low dimensional latent representations of complex data. The autoencoder framework consists of two mappings; the encoder which embeds information from the input space into a latent space, and a decoder which transforms point estimates from the latent space back into in the input space. Autoencoders have been successful in downstream tasks such as classification [@Creswell2017], outlier detection [@An2015], and data generation [@Wei2020]. 
 
-There exists many approaches for extending autoencoders to multi-view modelling. Generally, this involves learning separate encoder and decoder functions for each modality with the latent representations being combined or associated in some way. By far the most popular group of multi-view autoencoder models are multi-view extensions of Variational Autoencoders (VAEs) where the latent space is regularised by mapping the encoding distributions to a gaussian prior using a Kullback–Leibler (KL) divergence term. However, there are also other multi-view autoencoder frameworks, such as multi-view Adversarial Autoencoders (AAEs) [@Wang2019]. Here the latent space is regularised by mapping the encoding distribution to a prior (here a gaussian) using an auxiliary discriminator tasked with distinguishing samples from the posterior and prior distributions. The choice of AAE or VAE model may be influenced by various elements of the application process. For example, the encoding distribution which best describes the data or stability during training may impact the choice of model.
+There exist many software frameworks for extending autoencoders to multiple modalities. Generally, this involves learning separate encoder and decoder functions for each modality with the latent representations being combined or associated in some way. By far the most popular group of multi-view autoencoder models are multi-view extensions of Variational Autoencoders (VAEs) where the latent space is regularised by mapping the encoding distributions to a gaussian prior using a Kullback–Leibler (KL) divergence term. However, there are also other multi-view autoencoder frameworks, such as multi-view Adversarial Autoencoders (AAEs) [@Wang2019]. Here the latent space is regularised by mapping the encoding distribution to a prior (here a gaussian) using an auxiliary discriminator tasked with distinguishing samples from the posterior and prior distributions. The choice of AAE or VAE model may be influenced by various elements of the application process. For example, the encoding distribution which best describes the data or stability during training may impact the choice of model.
 
[…]
