+++ docs/JOSS/paper.md
@@ -90,4 +90,4 @@ Documentation is available (https://multi-view-ae.readthedocs.io/en/latest/) for
 # Acknowledgements
 We would like to thank Thomas Sutter, HyeongJoo Hwang and, Marco Lorenzi, and Brooks Paige, for their help understanding the Mixture-of-Product-of-Experts VAE [@Sutter2021] and mmJSD [@Sutter2021b], MVTCAE [@Hwang2021], mcVAE [@Antelmi2019], and MMVAE [@Shi2019] models, respectively.  
 
-ALA and NMB are supported by the EPSRC-funded UCL Centre for Doctoral Training in Intelligent, Integrated Imaging in Healthcare (i4health) and the Department of Health’s NIHR-funded Biomedical Research Centre at University College London Hospitals. AJ is supported by the Engineering Mathematics and Computing Lab (EMCL), Heidelberg University, the Helmholtz Association under the joint research school "HIDSS4Health – Helmholtz Information and Data Science School for Health", and the Heidelberg Institute for Theoretical Studies (HITS).
+ALA and NMB are supported by the EPSRC-funded UCL Centre for Doctoral Training in Intelligent, Integrated Imaging in Healthcare (i4health) and the Department of Health’s NIHR-funded Biomedical Research Centre at University College London Hospitals (EP/S021930/1). AJ is supported by the Engineering Mathematics and Computing Lab (EMCL), Heidelberg University, the Helmholtz Association under the joint research school "HIDSS4Health – Helmholtz Information and Data Science School for Health", and the Heidelberg Institute for Theoretical Studies (HITS).
