+++ docs/JOSS/paper.md
@@ -46,7 +46,7 @@ Even within these regularisation frameworks there are vast modelling differences
 
 Given the large number of multi-view autoencoders and versatility of architecture, it is important to consider which model would best suit the use case. `multi-view-AE` is a Python library which implements several variants of multi-view autoencoders in a simple, flexible, and easy-to-use framework. We would like to highlight the following benefits of our package. 
 
-Firstly, the `multi-view-AE` package is implemented with a similar interface to `scikit-learn` [@sklearn_api2013] with common and straight-forward functions implemented for all models. This makes it simple for users to train and evaluate models without requiring detailed knowledge of the methodology. Secondly, all models follow a modular structure. This gives users the flexibility to choose the class (such as the encoder or decoder network) from the available implementations, or to contribute their own. As such, the `multi-view-AE` package is accessible to both beginners, with off-the-shelf models, and experts, who wish to adapt the existing framework for further research purposes. Finally, the `multi-view-AE` package uses the `PyTorch-Lightning` API which offers the same functionality as raw `PyTorch` in a more structured and streamlined way. This offers users more flexibility, faster training and optimisation time, and high scalability.
+Firstly, the `multi-view-AE` package is implemented with a similar interface to `scikit-learn` [@sklearn_api2013] with common and straight-forward functions implemented for all models. This makes it simple for users to train and evaluate models without requiring detailed knowledge of the methodology. Secondly, all models follow a modular structure. This gives users the flexibility to choose the class (such as the encoder or decoder network) from the available implementations, or to contribute their own. As such, the `multi-view-AE` package is accessible to both beginners, with off-the-shelf models, and experts, who wish to adapt the existing framework for further research purposes. Finally, the `multi-view-AE` package uses the `PyTorch-Lightning` [@Falcon2019] API which offers the same functionality as raw `PyTorch` [@Paszke2019] in a more structured and streamlined way. This offers users more flexibility, faster training and optimisation time, and high scalability.
 
 # Statement of need
 Multi-view autoencoders have become a popular family of unsupervised learning methods in the field of multi-view learning. The flexibility of the form of the encoder and decoder functions, ease of extension to multiple views, generative properties, and adaptability to large scale datasets has contributed to the popularity of multi-view autoencoders compared to other multi-view methods. Subsequently, multi-view autoencoders have used to address challenges across a range of fields; such as anomaly detection from videos [@Deepak2021] or cross-modal generation of multi-omics data [@Minoura2021].
