+++ docs/JOSS/paper.bib
@@ -212,3 +212,21 @@ journal = {Circuits, Systems, and Signal Processing},
 	publisher = {Cold Spring Harbor Laboratory},
 	journal = {bioRxiv}
 }
+
+@incollection{Paszke2019,
+title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
+booktitle = {Advances in Neural Information Processing Systems 32},
+pages = {8024-8035},
+year = {2019},
+publisher = {Curran Associates, Inc.},
+}
+
+@article{Falcon2019,
+  title={Pytorch lightning},
+  author={Falcon, William and others},
+  journal={GitHub. Note: https://github. com/PyTorchLightning/pytorch-lightning},
+  volume={3},
+  number={6},
+  year={2019}
+}
