+++ docs/JOSS/paper.bib
@@ -4,6 +4,7 @@ booktitle = {Artificial Intelligence in the Age of Neural Networks and Brain Com
 publisher = {Academic Press},
 year = 2019,
 author = {{Serra}, A. and {Galdi}, P. and {Tagliaferri}, R.},
+doi = {10.1016/B978-0-12-815480-9.00013-X},
 }
 
 @article{Sjostrom1983,
@@ -13,6 +14,7 @@ volume = {150},
 pages = {61-70},
 year = {1983},
 author = {Michael Sjöström and Svante Wold and Walter Lindberg and Jan-Åke Persson and Harald Martens},
+doi = {https://doi.org/10.1016/S0003-2670(00)85460-4},
 }
 
 @article{Sadr2020,
@@ -22,7 +24,8 @@ author = {Michael Sjöström and Svante Wold and Walter Lindberg and Jan-Åke Pe
   year={2020},
   volume={8},
   month = {05},
-  pages={86984-86997}}
+  pages={86984-86997},
+  doi={10.1109/ACCESS.2020.2992063}}
 
 @article{Creswell2017,
 author = {Creswell, Antonia and Bharath, Anil},
@@ -48,7 +51,8 @@ author = {Wei, Ruoqi and Mahmood, Ausif},
 title = {Recent Advances in Variational Autoencoders With Representation Learning for Biomedical Informatics A Survey},
 year={2021},
 volume={9},
-pages={4939-4956}}
+pages={4939-4956},
+doi={10.1109/ACCESS.2020.3048309}}
 
 @article{Wang2019,
 title = {Adversarial correlated autoencoder for unsupervised multi-view representation learning},
@@ -56,7 +60,8 @@ journal = {Knowledge-Based Systems},
 volume = {168},
 pages = {109-120},
 year = {2019},
-author = {Xu Wang and Dezhong Peng and Peng Hu and Yongsheng Sang}
+author = {Xu Wang and Dezhong Peng and Peng Hu and Yongsheng Sang},
+doi = {https://doi.org/10.1016/j.knosys.2019.01.017},
 }
 
 @inproceedings{sklearn_api2013,
@@ -81,6 +86,7 @@ number = {5-6},
 pages = {261-278},
 year  = {2022},
 publisher = {Taylor & Francis},
+doi = {10.1080/01691864.2022.2035253},
 }
 
 @article{Suzuki2021,
@@ -176,7 +182,9 @@ journal = {Advances in Neural Information Processing Systems}
   number={7},
   pages={897-907},
   year={2000},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/09544120050135443},
+
 }
 
 @article{Krishnan2011,
@@ -187,6 +195,7 @@ pages = {455-75},
 title = {Partial Least Squares (PLS) Methods for Neuroimaging: A Tutorial and Review},
 volume = {56},
 journal = {NeuroImage},
+doi = {10.1016/j.neuroimage.2010.07.034},
 }
 
 @article{Minoura2021,
@@ -195,6 +204,7 @@ journal = {Cell Reports Methods},
 volume = {1},
 number = {5},
 year = {2021},
+doi = {https://doi.org/10.1016/j.crmeth.2021.100071},
 }
 
 @article{Deepak2021,
@@ -203,6 +213,7 @@ year = {2021},
 title = {Deep Multi-view Representation Learning for Video Anomaly Detection Using Spatiotemporal Autoencoders},
 volume = {40},
 journal = {Circuits, Systems, and Signal Processing},
+doi = {10.1007/s00034-020-01522-7},
 }
 
 @article {Minoura2021,
[…]
