+++ docs/JOSS/paper.bib
@@ -14,7 +14,7 @@ volume = {150},
 pages = {61-70},
 year = {1983},
 author = {Michael Sjöström and Svante Wold and Walter Lindberg and Jan-Åke Persson and Harald Martens},
-doi = {https://doi.org/10.1016/S0003-2670(00)85460-4},
+doi = {10.1016/S0003-2670(00)85460-4},
 }
 
 @article{Sadr2020,
@@ -61,7 +61,7 @@ volume = {168},
 pages = {109-120},
 year = {2019},
 author = {Xu Wang and Dezhong Peng and Peng Hu and Yongsheng Sang},
-doi = {https://doi.org/10.1016/j.knosys.2019.01.017},
+doi = {10.1016/j.knosys.2019.01.017},
 }
 
 @inproceedings{sklearn_api2013,
@@ -204,7 +204,7 @@ journal = {Cell Reports Methods},
 volume = {1},
 number = {5},
 year = {2021},
-doi = {https://doi.org/10.1016/j.crmeth.2021.100071},
+doi = {10.1016/j.crmeth.2021.100071},
 }
 
 @article{Deepak2021,
@@ -222,7 +222,7 @@ doi = {10.1007/s00034-020-01522-7},
 	year = {2021},
 	publisher = {Cold Spring Harbor Laboratory},
 	journal = {bioRxiv},
-  doi = {https://doi.org/10.1016/j.crmeth.2021.100071},
+  doi = {10.1016/j.crmeth.2021.100071},
 }
 
 @incollection{Paszke2019,
