+++ docs/JOSS/paper.bib
@@ -94,7 +94,8 @@ doi = {10.1080/01691864.2022.2035253},
   author={Masahiro Suzuki and Takaaki Kaneko and Yutaka Matsuo},
   journal={ArXiv},
   year={2021},
-  volume={abs/2107.13109}
+  volume={abs/2107.13109},
+  url       = {https://arxiv.org/abs/2107.13109},
 }
 
 @InProceedings{Antelmi2019,
@@ -106,6 +107,7 @@ doi = {10.1080/01691864.2022.2035253},
   volume = 	 {97},
   series = 	 {Proceedings of Machine Learning Research},
   publisher =    {PMLR},
+  url = 	 {https://proceedings.mlr.press/v97/antelmi19a.html},
 }
 
 @article{Wang2016,
@@ -162,7 +164,8 @@ journal={arXiv},
   author={Thomas M. Sutter and Imant Daunhawer and Julia E. Vogt},
   journal={ArXiv},
   year={2021},
-  volume={abs/2105.02470}
+  volume={abs/2105.02470},
+  url       = {https://arxiv.org/abs/2105.02470},
 }
 
 @article{Sutter2021b,
@@ -232,6 +235,7 @@ booktitle = {Advances in Neural Information Processing Systems 32},
 pages = {8024-8035},
 year = {2019},
 publisher = {Curran Associates, Inc.},
+url = {https://proceedings.neurips.cc/paper/2019/file/bdbca288fee7f92f2bfa9f7012727740-Paper.pdf},
 }
 
 @article{Falcon2019,
