### 2022-12-02 9e1aaa1
add missing url and dois
