+++ docs/JOSS/paper.bib
@@ -34,6 +34,7 @@ month = {03},
 title = {Denoising Adversarial Autoencoders},
 volume = {30},
 journal = {IEEE Transactions on Neural Networks and Learning Systems},
+url       = {http://arxiv.org/abs/1703.01220},
 }
 
 @article{An2015,
@@ -116,7 +117,8 @@ year = {2016},
 month = {10},
 title = {Deep Variational Canonical Correlation Analysis},
 journal={ArXiv},
-volume={abs/1610.03454}
+volume={abs/1610.03454},
+url       = {http://arxiv.org/abs/1610.03454},
 }
 
 @inproceedings{Shi2019,
@@ -125,7 +127,8 @@ author = {Shi, Yuge and Narayanaswamy, Siddharth and Paige, Brooks and Torr, Phi
 year = {2019},
 month = {11},
 pages = {},
-title = {Variational Mixture-of-Experts Autoencoders for Multi-Modal Deep Generative Models}
+title = {Variational Mixture-of-Experts Autoencoders for Multi-Modal Deep Generative Models},
+doi = {10.48550/ARXIV.1911.03393},
 }
 
 @inproceedings{Wu2018,
@@ -135,6 +138,7 @@ year = {2018},
 publisher = {Curran Associates Inc.},
 booktitle = {Proceedings of the 32nd International Conference on Neural Information Processing Systems},
 pages = {5580–5590},
+url       = {http://arxiv.org/abs/1802.05335},
 
 }
 
@@ -146,6 +150,7 @@ month = {11},
 pages = {},
 title = {Joint Multimodal Learning with Deep Generative Models},
 journal={arXiv}, 
+doi = {10.48550/ARXIV.1611.01891},
 }
 
 
@@ -156,7 +161,8 @@ journal={arXiv},
  publisher = {Curran Associates, Inc.},
  title = {Multi-View Representation Learning via Total Correlation Objective},
  volume = {34},
- year = {2021}
+ year = {2021},
+ url = {https://proceedings.neurips.cc/paper/2021/file/65a99bb7a3115fdede20da98b08a370f-Paper.pdf},
 }
 
 @article{Sutter2021,
@@ -174,7 +180,8 @@ year = {2021},
 month = {03},
 title = {Multimodal Generative Learning Utilizing Jensen-Shannon-Divergence},
 volume = {33},
-journal = {Advances in Neural Information Processing Systems}
+journal = {Advances in Neural Information Processing Systems},
+url       = {https://arxiv.org/abs/2006.08242},
 }
 
 @article{Cassel2000,
