### 2022-12-02 #4975 @alawryaguila
Hello! After doing a keyword search in the [this list of people](https://bit.ly/joss-reviewers), I can suggest the following reviewers (from the bottom of the list upwards):

- saran-nns
- abhi-glitchhg
- smith42
- thelinuxmaniac
- vs74

I do not know them personally, but I hope that they would find reviewing our software interesting.
