### 2022-12-29 350473b
update coordinated model in joss draft and readthedocs
