+++ docs/JOSS/paper.md
@@ -38,11 +38,11 @@ There exist many software frameworks for extending autoencoders to multiple moda
 
 ![Single view autoencoder frameworks; (a) vanilla autoencoder, (b) adversarial autoencoder, (c) variational autoencoder.](../figures/autoencoder_models.png)
 
-Even within these regularisation frameworks there are vast modelling differences to be considered when choosing the best model for the task at hand. \autoref{fig:AElatent} depicts two latent variable models for modelling two views of data; $X$ and $Y$. \autoref{fig:AElatent}a shows the joint latent variable model [@Suzuki2022] where both views, $X$ and $Y$, share an underlying factor. The latent variable model in \autoref{fig:AElatent}b shows a coordinated model [@Suzuki2022], which assumes some relationship between the latent variables, $z_x$ and $z_y$ of $X$ and $Y$ respectively. Which latent variable model is most appropriate depends on the desired outcome of the learning task. Example multi-view autoencoder frameworks built for these two latent variable models are given in \autoref{fig:AEexample}.
+Even within these regularisation frameworks there are vast modelling differences to be considered when choosing the best model for the task at hand. \autoref{fig:AElatent} depicts two latent variable models for modelling two views of data; $X$ and $Y$. \autoref{fig:AElatent}a shows the joint latent variable model [@Suzuki2022] where both views, $X$ and $Y$, share an underlying factor. The latent variable model in \autoref{fig:AElatent}b shows a coordinated model, which assumes some relationship between the latent variables, $z_x$ and $z_y$ of $X$ and $Y$ respectively. Which latent variable model is most appropriate depends on the desired outcome of the learning task. Example multi-view autoencoder frameworks built for these two latent variable models are given in \autoref{fig:AEexample}.
 
 ![Latent variable models for two input views. Latent variable model where data $X$ and $Y$ (a) share an underlying latent factor $z$ (b) have associated latent factors $z_x$ and $z_y$.\label{fig:AElatent}](../figures/latent_models.png)
 
-![Example frameworks of a two-view autoencoder for data $X$ and $Y$ for a (a) joint model, where the individual latent spaces are combined and the reconstruction is carried out from the joint latent space, and a (b) coordinated model [@Suzuki2022], where the latent representations are coordinated either by cross view generation or an addition loss term for association between the latent variables.\label{fig:AEexample}](../figures/joint_coordinated_models.png)
+![Example frameworks of a two-view autoencoder for data $X$ and $Y$ for a (a) joint model, where the individual latent spaces are combined and the reconstruction is carried out from the joint latent space, and a (b) coordinated model, where the latent representations are coordinated either by cross view generation or an addition loss term for association between the latent variables.\label{fig:AEexample}](../figures/joint_coordinated_models.png)
 
 Given the large number of multi-view autoencoders and versatility of architecture, it is important to consider which model would best suit the use case. `multi-view-AE` is a Python library which implements several variants of multi-view autoencoders in a simple, flexible, and easy-to-use framework. We would like to highlight the following benefits of our package. 
 
