+++ docs/JOSS/paper.md
@@ -38,7 +38,7 @@ There exist many software frameworks for extending autoencoders to multiple moda
 
 ![Single view autoencoder frameworks; (a) vanilla autoencoder, (b) adversarial autoencoder, (c) variational autoencoder.](../figures/autoencoder_models.png)
 
-Even within these regularisation frameworks there are vast modelling differences to be considered when choosing the best model for the task at hand. \autoref{fig:AElatent} depicts two latent variable models for modelling two views of data; $X$ and $Y$. \autoref{fig:AElatent}a shows the joint latent variable model [@Suzuki2022] where both views, $X$ and $Y$, share an underlying factor. The latent variable model in \autoref{fig:AElatent}b shows a coordinated model, which assumes some relationship between the latent variables, $z_x$ and $z_y$ of $X$ and $Y$ respectively. Which latent variable model is most appropriate depends on the desired outcome of the learning task. Example multi-view autoencoder frameworks built for these two latent variable models are given in \autoref{fig:AEexample}.
+Even within these regularisation frameworks there are vast modelling differences to be considered when choosing the best model for the task at hand. \autoref{fig:AElatent} depicts two possible latent variable models for modelling two views of data; $X$ and $Y$. \autoref{fig:AElatent}a shows the joint latent variable model [@Suzuki2022] where both views, $X$ and $Y$, share an underlying factor. The latent variable model in \autoref{fig:AElatent}b shows a coordinated model, which assumes some relationship between the latent variables, $z_x$ and $z_y$ of $X$ and $Y$ respectively. Which latent variable model is most appropriate depends on the desired outcome of the learning task. Example multi-view autoencoder frameworks built for these two latent variable models are given in \autoref{fig:AEexample}.
 
 ![Latent variable models for two input views. Latent variable model where data $X$ and $Y$ (a) share an underlying latent factor $z$ (b) have associated latent factors $z_x$ and $z_y$.\label{fig:AElatent}](../figures/latent_models.png)
 
