### 2023-01-11 #4975 @arfon
@smith42 @thelinuxmaniac @vs74 :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is Multi-view-AE: A Python package for multi-view autoencoder models: https://github.com/alawryaguila/multi-view-AE

Many thanks
Arfon
