### 2023-01-22 #4975 @arfon
@saran-nns @abhi-glitchhg 👋 would you be willing to review this submission for JOSS? The submission under consideration is Multi-view-AE: A Python package for multi-view autoencoder models: https://github.com/alawryaguila/multi-view-AE

Many thanks
Arfon
