### 2023-01-30 alawryaguila/multi-view-AE#22 @Saran-nns
[REVIEW]: Multi-view-AE: A Python package for multi-view autoencoder models
Hi @alawryaguila @ajayme
I have few suggestions and recommendations regarding the package and the paper,

### API:
- [ ] I have a concern about the config file settings. Your documentation references a "configuration yaml" file which lives inside the installed python package (or if not where it should be located and how it should be defined) and is suggested to be edited by the user. This doesn't seem like a good solution for configuring the core part of your library, as the site package directory may not always be editable by the user (e.g. if the library was installed by an administrator, the library is shared across users, etc.)

I would suggest changing the way this is handled. Since the args defined by the user can override the defaults in the config, I think you could have these as the default params/values inside the model class itself. Also I recommend to include the default params/values in the documentation.

### Documentation:
- [ ] Usage example in the documentation breaks;
“datasets and transforms” are not imported
- [ ] MissingConfigException: Cannot find primary config './config_folder/test_config.yaml'. Check that it's in your config search path.

### Tests:
- [ ] I think it would be great to add some documentation to the tests, e.g., inline comments about what the tests are doing and why / why not they should fail.
### Authorship and contributions:
- [ ] We see only 2 authors contributed to the package, could you please explain how others contributed to the work? May be, it is appropriate to include a contribution statement in the paper, @arfon what do you think?
