### 2023-01-30 alawryaguila/multi-view-AE#22 @arfon
> We see only 2 authors contributed to the package, could you please explain how others contributed to the work? May be, it is appropriate to include a contribution statement in the paper, @arfon what do you think?

This is quite common, especially for JOSS papers. Note, not all authors have to have contributed to the code: https://joss.readthedocs.io/en/latest/review_criteria.html?highlight=authorship#authorship
