### 2023-01-30 alawryaguila/multi-view-AE#22 @alawryaguila
Hi @Saran-nns, thank you for your comments and suggests for our package and documentation. We will begin working on your suggestions and update this issue once we believe we have completed the changes. Please find our comments below:

> I have a concern about the config file settings. Your documentation references a "configuration yaml" file which lives inside the installed python package (or if not where it should be located and how it should be defined) and is suggested to be edited by the user. This doesn't seem like a good solution for configuring the core part of your library, as the site package directory may not always be editable by the user (e.g. if the library was installed by an administrator, the library is shared across users, etc.)

Whilst the default configuration file is contained within the installed python package, this need not be the case for the configuration file provided by the user. The user can provide the absolute or relative path to a configuration file on their machine. Any parameters set in the user configuration file will override the default package parameters for the instance of the model for which the configuration file was used. The user doesn't need to have access to the package directory and the user configuration file will not override the default parameters for other users. Please let us know if this answers your concerns.

The default parameters for each model are contained within model specific yaml files (in the configs/model_type) directory. The default parameters saved to the relevant model class when initialised. These files are not designed to be edited by the user rather for the user to view the model specific default parameters.

> Also I recommend to include the default params/values in the documentation.

Thank you for your comment. We will update the documentation accordingly.

>  Usage example in the documentation breaks;
“datasets and transforms” are not imported
 MissingConfigException: Cannot find primary config './config_folder/test_config.yaml'. Check that it's in your config search path. Must be related to API issue above.
 I tested the package with python 3.10 and it works flawlessly. Good to update python compatibility <=3.10

Noted thank you. We will fix these issues.

> I think it would be great to add some documentation to the tests, e.g., inline comments about what the tests are doing and why / why not they should fail.

Noted thank you. We will add documentation to the tests.

> We see only 2 authors contributed to the package, could you please explain how others contributed to the work?

The remaining authors contributed towards the code design, model theory, and paper and did not contribute directly to the code.

> Contribution guidelines are missing. It would be convenient to have it in the readme as well as in the documentation.

Noted thank you. We will update.
