+++ README.md
@@ -9,6 +9,7 @@ The ``multi-view-AE`` library is compatible, and has been tested, with the follo
 * 3.7
 * 3.8
 * 3.9
+* 3.10
 
 ### Installation
 Clone this repository and move to folder:
