### 2023-02-03 f521b38
add contribution information and fix running on gpu
