+++ docs/source/developer_info/contributions.rst
@@ -1,0 +1,60 @@
+How to contribute
+===========
+
+Contributions to the ``multiviewAE`` library are welcome and greatly appreciated! Here are some ways you can contribute:
+
+Reporting a bug
+----------
+
+Please report bugs by submitting an issue at https://github.com/alawryaguila/multi-view-AE/issues
+
+If you are reporting a bug, please include the following information:
+
+* A quick summary and/or background.
+* Your operating system name and version.
+* Details about your local setup that might be helpful in troubleshooting e.g. python version, library versions
+* Detailed steps to reproduce the bug.
+* What you expected to happen.
+* What actually happens.
+
+Proposing a new feature
+----------
+
+The best way to propose a new feature is by submitting an issue https://github.com/alawryaguila/multi-view-AE/issues
+
+To propose a feature please include:
+
+* Describe in detail how the new feature would work.
+* Explain the use case of the new feature.
+* Please keep the scope as narrow and specific as possible, to make it easier to implement.
+
+Writing documentation
+----------
+
+The latest documentation for the ``multiviewAE`` library is available at https://multi-view-ae.readthedocs.io/en/latest/
+
+If any documentation is unclear or requires correction, please submit an issue.
+
+Submitting changes
+----------
+To submit your code when fixing bugs, documentation, or implementing new features, please follow the steps below.
+
+1. Fork the ``multiviewAE`` repository on GitHub.
+
+2. Clone your fork locally::
+
+    $ git clone git@github.com:your_name_here/multi-view-AE.git
+
+3. Create a branch for local development::
+
+    $ git checkout -b name-of-your-bugfix-or-feature
+
+    Make your desired changes on your local branch.
+
+4. Commit your changes and push your branch to GitHub::
+
+    $ git add .
+    $ git commit -m "Your detailed description of your changes."
+    $ git push origin name-of-your-bugfix-or-feature
+
+5. Submit a pull request through GitHub. 
