### 2023-02-03 e488e0d
update bullet points in documentation
