### 2023-02-03 46ed24b
update bullet points in documentation
