### 2023-02-03 9a91a59
update bullet points in documentation
