+++ docs/requirements.txt
@@ -1,4 +1,5 @@
 protobuf==3.20.1
+docutils<0.17
 sphinx
 sphinx-autodoc-typehints
 sphinx-gallery
