### 2023-02-08 a1730ba
fix stability of probability pooling functions
