### 2023-02-13 alawryaguila/multi-view-AE#22 @Saran-nns
Thanks for considering my suggestions @alawryaguila. I am convinced with the config file settings now.

Also, i found a small correction in the paper line #14 . ```Autoencoders are unsupervised generative models```.  Sorry that I have different view on this. Autoencoders are typically unsupervised algorithms for creating a latent representation from input-output mapping. But, when the latent representation is trained to maximize the probability of the input data, the auto encoder becomes a generative model. You agree? I think we should clarify this in the paper.

Thank you :)
