### 2023-02-13 ead9524
removed Sphinx warnings. added pages for default and model params (from yaml files)
