+++ multiviewae/models/mvtcae.py
@@ -14,17 +14,18 @@ class mvtCAE(BaseModelVAE):
     NOTE: This implementation currently only caters for a PoE posterior distribution. MoE and MoPoE posteriors will be included in further work.
 
     Args:
-    cfg (str): Path to configuration file. Model specific parameters in addition to default parameters:
-        model.beta (int, float): KL divergence weighting term.
-        model.alpha (int, float): Log likelihood, Conditional VIB and VIB weighting term.
-        encoder.default._target_ (multiae.architectures.mlp.VariationalEncoder): Type of encoder class to use.
-        encoder.default.enc_dist._target_ (multiae.base.distributions.Normal, multiae.base.distributions.MultivariateNormal): Encoding distribution.
-        decoder.default._target_ (multiae.architectures.mlp.VariationalDecoder): Type of decoder class to use.
-        decoder.default.init_logvar(int, float): Initial value for log variance of decoder.
-        decoder.default.dec_dist._target_ (multiae.base.distributions.Normal, multiae.base.distributions.MultivariateNormal): Decoding distribution.
-
-    input_dim (list): Dimensionality of the input data.
-    z_dim (int): Number of latent dimensions.
+        cfg (str): Path to configuration file. Model specific parameters in addition to default parameters:
+            
+            - model.beta (int, float): KL divergence weighting term.
+            - model.alpha (int, float): Log likelihood, Conditional VIB and VIB weighting term.
+            - encoder.default._target_ (multiae.architectures.mlp.VariationalEncoder): Type of encoder class to use.
+            - encoder.default.enc_dist._target_ (multiae.base.distributions.Normal, multiae.base.distributions.MultivariateNormal): Encoding distribution.
+            - decoder.default._target_ (multiae.architectures.mlp.VariationalDecoder): Type of decoder class to use.
+            - decoder.default.init_logvar(int, float): Initial value for log variance of decoder.
+            - decoder.default.dec_dist._target_ (multiae.base.distributions.Normal, multiae.base.distributions.MultivariateNormal): Decoding distribution.
+    
+        input_dim (list): Dimensionality of the input data.
+        z_dim (int): Number of latent dimensions.
 
     References
     ----------
