+++ docs/source/api/models.rst
@@ -1,16 +1,6 @@
 Models
 ==================
 
-.. contents:: Contents
-    :local:
-
-.. currentmodule:: multiviewae.models
-.. autosummary::
-   :nosignatures:
-
-    {% for cls in multiviewae.models.classes %}
-    {{ cls }}
-    {% endfor %}
-
 .. automodule:: multiviewae.models
-   :members:
+    :members:
+    

[…]
