### 2023-02-13 alawryaguila/multi-view-AE#22 @alawryaguila
Hi @Saran-nns thank you for your comment and bringing this correction to our attention. We agree that the characterization of all autoencoders as generative models is incorrect (only the adversarial and variational models can be considered as generative models but not vanilla autoencoders) and we will update the paper. Thank you!
