+++ docs/source/api/parameters.rst
@@ -4,7 +4,7 @@ Parameters and Configurables
 General and Default
 ---------------------
 
-.. datatemplate:yaml:: multiviewae.configs.default.yaml
+.. datatemplate:yaml:: ../../../multiviewae/configs/default.yaml
    :template: default_params.tmpl
 
 Model-specific
