### 2023-02-15 alawryaguila/multi-view-AE#22 @Saran-nns
@alawryaguila thank you again. I tried the example [usage](https://multi-view-ae.readthedocs.io/en/latest/documentation/user_guide.html#model-fit) again and it failed while calling `fit`. Please see the colab notebook [here](https://colab.research.google.com/drive/1mFEZ-oo8_8ezWKjwTjP_HWhxlgjz1E4V?usp=sharing) for the traceback.
