### 2023-02-15 f209650
Update exampleMNIST.py
