### 2023-02-15 3a48864
Update exampleMNIST.py
