+++ examples/exampleMNIST.py
@@ -105,8 +105,8 @@ axarr[1].imshow(pred_sample.reshape(28,14))
 #Reconstruction plots - how well can the VAE do cross view reconstruction?
 
 data_sample = data_1[20]
-#indices: view 2 latent, view 2 decoder, sample 21
-pred_sample = pred[1][1][20]
+#indices: view 1 latent, view 2 decoder, sample 21
+pred_sample = pred[0][1][20]
 
 fig, axarr = plt.subplots(1, 2)
 plt.rc('text', usetex=True)
