### 2023-02-15 7b69de4
Update user_guide.rst
