### 2023-02-15 alawryaguila/multi-view-AE#22 @alawryaguila
Hi @Saran-nns  apologies for this, I had not noticed the discrepancy between the example code in the documentation and the example scripts in the ./examples folder. This issue is now fixed, please let me know if there are any other issues.
