+++ docs/source/documentation/user_guide.rst
@@ -4,7 +4,7 @@ User Guide
 User guide for initialising and running models from the ``multiviewAE`` library. 
 
 Initialise model 
-----------------
+----------
 
 .. code-block:: python
 
@@ -21,11 +21,11 @@ Initialise model
    data_2 = MNIST_1.train_data[:, :, 14:].reshape(-1,392).float()/255.
 
    mvae = mVAE(cfg="./examples/config/example_mnist.yaml",
-        input_dim=[784, 784],
+        input_dim=[392, 392],
         z_dim=64)
 
    mcvae = mcVAE(cfg="./examples/config/example_mnist.yaml",
-        input_dim=[784, 784],
+        input_dim=[392, 392],
         z_dim=64)
 
 
@@ -42,7 +42,7 @@ Model fit
 When fitting the model, the user must provide input each view of the training data. The user can optionally provide the ``max_epochs`` and ``batch_size``. These would override the settings in the configuration file. 
 
 Model predictions
------------------
+----------
 
 We can use a trained model to predict the latent dimensions or reconstructions. The structure of the latent and reconstruction list will depend on the type of model. Below shows an example for joint, ``MVAE``,  and coordinate, ``mcVAE``, multi-view VAE models.
 
@@ -54,9 +54,10 @@ We can use a trained model to predict the latent dimensions or reconstructions. 
    data_test_1 = MNIST_1.test_data[:, :, :14].reshape(-1,392).float()/255.
    data_test_2 = MNIST_1.test_data[:, :, 14:].reshape(-1,392).float()/255.
 
-   mvae_latent = mvae.predict_latent(data_test_1, data_test_2)
+   mvae_latent = mvae.predict_latents(data_test_1, data_test_2)
 
-   mcvae_latent = mcvae.predict_latent(data_test_1, data_test_2)
+   mcvae_latent = mcvae.predict_latents(data_test_1, data_test_2)
+
    mcvae_latent_view1, mcvae_latent_view2 = mcvae_latent[0], mcvae_latent[1]
 
    mvae_reconstruction = mvae.predict_reconstruction(data_test_1, data_test_2)
@@ -72,9 +73,41 @@ We can use a trained model to predict the latent dimensions or reconstructions. 
    mcvae_reconstruction_view1_latent2 = mcvae_reconstruction[1][0] #view 1 reconstruction from latent 2
    mcvae_reconstruction_view2_latent2 = mcvae_reconstruction[1][1] #view 2 reconstruction from latent 2
 
-
-Model loading
+Model results
 -------------
+
+We can explore the model results, for example using ``matplotlib``. 
+
+.. code-block:: python
+
+   import matplotlib.pyplot as plt #NOTE: matplotlib is not installed with the library and must be installed separately 
+   
+   #Reconstruction plots - how well can the VAE do same view reconstruction?
+
+   data_sample = data_test_1[20]
+   #indices: view 1 latent, view 1 decoder, sample 21
+   pred_sample = mcvae_reconstruction_view1_latent1[20]
+
+   fig, axarr = plt.subplots(1, 2)
+   axarr[0].imshow(data_sample.reshape(28,14))
+   axarr[1].imshow(pred_sample.reshape(28,14))
+   plt.show()
+   plt.close()
+   
+   #Reconstruction plots - how well can the VAE do cross view reconstruction?
+
+   #indices: view 1 latent, view 2 decoder, sample 21
+   data_sample = data_test_2[20]
+   pred_sample = mcvae_reconstruction_view2_latent1[20]
+
+   fig, axarr = plt.subplots(1, 2)
+   axarr[0].imshow(data_sample.reshape(28,14))
+   axarr[1].imshow(pred_sample.reshape(28,14))
+   plt.show()
[…]
