### 2023-02-15 c04e9cf
Update example_mnist.yaml
