+++ examples/config/example_mnist.yaml
@@ -3,7 +3,8 @@
 model:
   z_dim: 64
   s_dim: 30
-
+  learning_rate: 0.00001
+  
 encoder:
   default:
     hidden_layer_dim: [256, 256]
