### 2023-02-15 efd06fb
Update exampleMNIST.py
