### 2023-02-15 alawryaguila/multi-view-AE#22 @Saran-nns
thank you @alawryaguila, the usage example works fine now and also a small reminder to update the paper (line #14)
