### 2023-02-15 alawryaguila/multi-view-AE#22 @alawryaguila
thank you @Saran-nns  that's great. The paper should be updated in the joss branch of the library. Please let me know if you are unable to see the updated draft.
