### 2023-02-15 4534486
Update exampleMNIST.py
