+++ examples/exampleMNIST.py
@@ -26,7 +26,7 @@ target_test = MNIST_1.test_labels.numpy()
 
 #Define parameters
 input_dim=[392,392]
-max_epochs = 10
+max_epochs = 500
 batch_size = 2000
 latent_dim = 2
 
@@ -97,11 +97,10 @@ data_sample = data_1[20]
 pred_sample = pred[0][0][20]
 
 fig, axarr = plt.subplots(1, 2)
-plt.rc('text', usetex=True)
-plt.rc('font', family='serif')
 axarr[0].imshow(data_sample.reshape(28,14))
 axarr[1].imshow(pred_sample.reshape(28,14))
-
+plt.show()
+plt.close()
 #Reconstruction plots - how well can the VAE do cross view reconstruction?
 
 data_sample = data_1[20]
@@ -109,7 +108,7 @@ data_sample = data_1[20]
 pred_sample = pred[1][1][20]
 
 fig, axarr = plt.subplots(1, 2)
-plt.rc('text', usetex=True)
-plt.rc('font', family='serif')
 axarr[0].imshow(data_sample.reshape(28,14))
 axarr[1].imshow(pred_sample.reshape(28,14))
+plt.show()
+plt.close()
