+++ docs/source/api/models.rst
@@ -1,6 +1,0 @@
-Models
-==================
-
-.. automodule:: multiviewae.models
-    :members:
-    
