+++ docs/source/index.rst
@@ -1,40 +1,0 @@
-Welcome to Multi-view-AE's documentation!
-=========================================
-
-**Multi-view-AE** is a Python library for multi-view autoencoder models. Please see the user guides below.
-
---------
-
-Documentation
-=============
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Using multi-view-AE
-   
-   documentation/implemented_models
-   documentation/user_guide
-   documentation/model_frameworks
-   documentation/parameter_configurations
-   documentation/limitations
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Reference
-
-   api/models
-   api/architectures
-   api/parameters
-      
-.. toctree::
-   :maxdepth: 1
-   :caption: Developer information
-
-   developer_info/contributions
-   developer_info/support
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`search`

[…]
