### 2023-03-18 #5093 @abhi-glitchhg
Hey @alawryaguila , Sorry for the delay.

The installation instructions are missing from the  [website](https://multi-view-ae.readthedocs.io/en/latest/documentation/user_guide.html). Having an installation guide page in the documentation would be more user-friendly. Thoughts?

Also how about adding the project to [Pypi](https://pypi.org/) (This is not a blocker, just a suggestion)
