+++ docs/source/documentation/installation.rst
@@ -1,0 +1,28 @@
+Installation
+===========
+
+Installation instructions for the ``multiviewAE`` library. 
+
+Clone the below repository and move to folder:
+
+.. code-block:: python
+  git clone https://github.com/alawryaguila/multi-view-AE
+  cd multi-view-AE
+
+
+Create the customised python environment:
+
+.. code-block:: python
+  conda create --name mvae python=3.9
+
+
+Activate python environment:
+
+.. code-block:: python
+  conda activate mvae
+
+
+Install the ``multi-view-AE`` package:
+
+.. code-block:: python
+  pip install ./
