### 2023-03-20 #5093 @alawryaguila
> Hey @alawryaguila , Sorry for the delay.
>
> The installation instructions are missing from the [website](https://multi-view-ae.readthedocs.io/en/latest/documentation/user_guide.html). Having an installation guide page in the documentation website would be more user-friendly. Thoughts?
>
> Also how about adding the project to [Pypi](https://pypi.org/) (This is not a blocker, just a suggestion)

Thank you for your comments @abhi-glitchhg, we have now added the installation instructions to the readthedocs link. Great suggestion about adding the project to Pypi. We had originally considered adding to Pypi prior to submission, so we will make sure to do this over the next few weeks.
