+++ tests/user_config/mmjsd.yaml
@@ -1,7 +1,6 @@
 # @package _global_
 
 model:
-  use_GPU: True
   seed: 1000
 
   z_dim: 12
