+++ docs/JOSS/paper.bib
@@ -1,0 +1,255 @@
+@inbook{Serra2018,
+title = {Multiview Learning in Biomedical Applications},
+booktitle = {Artificial Intelligence in the Age of Neural Networks and Brain Computing},
+publisher = {Academic Press},
+year = 2019,
+author = {{Serra}, A. and {Galdi}, P. and {Tagliaferri}, R.},
+doi = {10.1016/B978-0-12-815480-9.00013-X},
+}
+
+@article{Sjostrom1983,
+title = {A multivariate calibration problem in analytical chemistry solved by partial least-squares models in latent variables},
+journal = {Analytica Chimica Acta},
+volume = {150},
+pages = {61-70},
+year = {1983},
+author = {Michael Sjöström and Svante Wold and Walter Lindberg and Jan-Åke Persson and Harald Martens},
+doi = {10.1016/S0003-2670(00)85460-4},
+}
+
+@article{Sadr2020,
+  author={Sadr, Hossein and Pedram, Mir Mohsen and Teshnehlab, Mohammad},
+  journal={IEEE Access}, 
+  title={Multi-View Deep Network: A Deep Model Based on Learning Features From Heterogeneous Neural Networks for Sentiment Analysis}, 
+  year={2020},
+  volume={8},
+  month = {05},
+  pages={86984-86997},
+  doi={10.1109/ACCESS.2020.2992063}}
+
+@article{Creswell2017,
+author = {Creswell, Antonia and Bharath, Anil},
+year = {2017},
+month = {03},
+title = {Denoising Adversarial Autoencoders},
+volume = {30},
+journal = {IEEE Transactions on Neural Networks and Learning Systems},
+url       = {http://arxiv.org/abs/1703.01220},
+}
+
+@article{An2015,
+  title={Variational autoencoder based anomaly detection using reconstruction probability},
+  author={An, Jinwon and Cho, Sungzoon},
+  journal={Special Lecture on IE},
+  volume={2},
+  number={1},
+  pages={1-18},
+  year={2015}
+}
+
+@article{Wei2020,
+author = {Wei, Ruoqi and Mahmood, Ausif},
+title = {Recent Advances in Variational Autoencoders With Representation Learning for Biomedical Informatics A Survey},
+year={2021},
+volume={9},
+pages={4939-4956},
+doi={10.1109/ACCESS.2020.3048309}}
+
+@article{Wang2019,
+title = {Adversarial correlated autoencoder for unsupervised multi-view representation learning},
+journal = {Knowledge-Based Systems},
+volume = {168},
+pages = {109-120},
+year = {2019},
+author = {Xu Wang and Dezhong Peng and Peng Hu and Yongsheng Sang},
+doi = {10.1016/j.knosys.2019.01.017},
+}
+
+@inproceedings{sklearn_api2013,
+  author    = {Lars Buitinck and Gilles Louppe and Mathieu Blondel and
+               Fabian Pedregosa and Andreas Mueller and Olivier Grisel and
+               Vlad Niculae and Peter Prettenhofer and Alexandre Gramfort
+               and Jaques Grobler and Robert Layton and Jake VanderPlas and
+               Arnaud Joly and Brian Holt and Ga{\"{e}}l Varoquaux},
+  title     = {{API} design for machine learning software: experiences from the scikit-learn
+               project},
+  booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
+  year      = {2013},
+  pages = {108-122},
+}
+
+@article{Suzuki2022,
+author = {Masahiro Suzuki and Yutaka Matsuo},
[…]
