+++ docs/JOSS/paper.md
@@ -1,0 +1,93 @@
+---
+title: 'Multi-view-AE: A Python package for multi-view autoencoder models'
+tags:
+  - Python
+  - Autoencoders
+  - Multi-view
+  - Unsupervised learning
+  - Representation learning
+  - Data generation
+authors:
+  - name: Ana Lawry Aguila
+    orchid: 0000-0003-0727-3274
+    corresponding: true
+    affiliation: 1
+  - name: Alejandra Jayme
+    affiliation: 2
+  - name: Nina Montaña-Brown
+    affiliation: 1
+  - name: Vincent Heuveline
+    affiliation: 2
+  - name: Andre Altmann
+    orchid: 0000-0002-9265-2393
+    affiliation: 1
+
+affiliations:
+  - name: Centre for Medical Image Computing (CMIC), Medical Physics and Biomedical Engineering, University College London (UCL), London, UK
+    index: 1
+  - name: Engineering Mathematics and Computing Lab (EMCL), Heidelberg, Germany
+    index: 2
+date: 25 November 2022
+bibliography: paper.bib
+---
+Often, data can be naturally described via multiple views or modalities. For example, we could consider an image and the corresponding text as different modalities. These modalities contain complementary information which can be modelled jointly using multi-view methods. The joint modelling of multiple modalities has been explored in many research fields such as medical imaging [@Serra2018], chemistry [@Sjostrom1983], and natural language processing [@Sadr2020]. 
+
+Autoencoders are unsupervised models which learn low dimensional latent representations of complex data. The autoencoder framework consists of two mappings; the encoder which embeds information from the input space into a latent space, and a decoder which transforms point estimates from the latent space back into in the input space. Autoencoders have been successful in downstream tasks such as classification [@Creswell2017], outlier detection [@An2015], and data generation [@Wei2020]. 
+
+There exist many software frameworks for extending autoencoders to multiple modalities. Generally, this involves learning separate encoder and decoder functions for each modality with the latent representations being combined or associated in some way. By far the most popular group of multi-view autoencoder models are multi-view extensions of Variational Autoencoders (VAEs) where the latent space is regularised by mapping the encoding distributions to a gaussian prior using a Kullback–Leibler (KL) divergence term. However, there are also other multi-view autoencoder frameworks, such as multi-view Adversarial Autoencoders (AAEs) [@Wang2019]. Here the latent space is regularised by mapping the encoding distribution to a prior (here a gaussian) using an auxiliary discriminator tasked with distinguishing samples from the posterior and prior distributions. The choice of AAE or VAE model may be influenced by various elements of the application process. For example, the encoding distribution which best describes the data or stability during training may impact the choice of model.
+
+![Single view autoencoder frameworks; (a) vanilla autoencoder, (b) adversarial autoencoder, (c) variational autoencoder.](../figures/autoencoder_models.png)
+
+Even within these regularisation frameworks there are vast modelling differences to be considered when choosing the best model for the task at hand. \autoref{fig:AElatent} depicts two possible latent variable models for modelling two views of data; $X$ and $Y$. \autoref{fig:AElatent}a shows the joint latent variable model [@Suzuki2022] where both views, $X$ and $Y$, share an underlying factor. The latent variable model in \autoref{fig:AElatent}b shows a coordinated model, which assumes some relationship between the latent variables, $z_x$ and $z_y$ of $X$ and $Y$ respectively. Which latent variable model is most appropriate depends on the desired outcome of the learning task. Example multi-view autoencoder frameworks built for these two latent variable models are given in \autoref{fig:AEexample}.
+
[…]
