+++ docs/source/api/models.rst
@@ -1,0 +1,6 @@
+Models
+==================
+
+.. automodule:: multiviewae.models
+    :members:
+    

[…]
