+++ multi-view-ae/base/validation.py
@@ -1,189 +1,0 @@
-from schema import Schema, And, Or, Optional, SchemaError, Regex
-
-SUPPORTED_ENCODERS = [
-            "mlp.Encoder",
-            "mlp.VariationalEncoder",
-            "cnn.Encoder",
-            "cnn.VariationalEncoder"
-        ]
-
-SUPPORTED_DECODERS = [
-            "mlp.Decoder",
-            "mlp.VariationalDecoder",
-            "cnn.Decoder"
-        ]
-
-SUPPORTED_DISCRIMINATORS = [
-            "mlp.Discriminator"
-        ]
-
-SUPPORTED_DISTRIBUTIONS = [
-            "distributions.Default",
-            "distributions.Normal",
-            "distributions.MultivariateNormal",
-            "distributions.Bernoulli",
-            "distributions.Laplace"
-        ]
-
-SUPPORTED_JOIN = [
-            "PoE",
-            "Mean"
-        ]
-
-UNSUPPORTED_ENC_DIST = [
-            "distributions.Bernoulli"
-        ]
-
-UNSUPPORTED_PRIOR_DIST = [
-            "distributions.Default",
-            "distributions.Bernoulli"
-        ]
-
-def return_or(params=[], msg="invalid"):
-    assert(len(params) > 0)
-
-    p = ', '.join('"{0}"'.format(str(w)) for w in params)
-    return f"Or({p}, error='{msg}')"
-
-def return_regexor(params=[], msg="invalid"):
-    assert(len(params) > 0)
-    p = '|'.join('{0}'.format(str(w)) for w in params)
-    return f"Regex(r'(.*?).({p})$', error='{msg}')"
-
-def list_sub(a, b):
-    return list(set(a) - set(b))
-
-config_schema = Schema({
-    "model": {
-        "use_GPU": bool,
-        "save_model": bool,
-        "seed_everything": bool,
-        "seed": And(int, lambda x: 0 <= x <= 4294967295),
-        "z_dim": int,
-        "learning_rate": And(float, lambda x: 0 < x < 1),
-        "sparse": bool,
-        "threshold": Or(And(float, lambda x: 0 < x < 1), 0),
-        Optional("eps"): And(float, lambda x: 0 < x <= 1e-10),
-        Optional("beta"): And(Or(int, float), lambda x: x > 0),
-        Optional("K"): And(int, lambda x: x >= 1),
-        Optional("alpha"): And(float, lambda x: x > 0),
-        Optional("private"): bool,
-        Optional("join_type"): eval(return_or(params=SUPPORTED_JOIN,
-                        msg="model.join_type: unsupported or invalid join type"))
-    },
-    "datamodule": {
-        "_target_": "multiae.base.dataloaders.MultiviewDataModule",    #TODO: how about user-defined classes?
-        "batch_size": Or(And(int, lambda x: x > 0), None),
-        "is_validate": bool,
-        "train_size": And(float, lambda x: 0 < x < 1)
-    },
-    "encoder": {
-        "default": {
-            "_target_" : eval(return_regexor(params=SUPPORTED_ENCODERS,
-                            msg="encoder._target_: unsupported or invalid encoder")),
-            Optional("hidden_layer_dim"): [And(int, lambda x: x > 0)],
-            Optional(Regex(r'^layer\d$')) : {
-                "layer": str, # TODO: fix this. should be specific torch.nn layers
-                # TODO: how about parameters of each layer? will raise error anyway...
-            },
-            "bias": bool,
-            "non_linear": bool,
-            "enc_dist": {
-                    "_target_": eval(return_regexor(params=list_sub(SUPPORTED_DISTRIBUTIONS, UNSUPPORTED_ENC_DIST),
[…]
