+++ multi-view-ae/models/me_mvae.py
@@ -1,198 +1,0 @@
-import torch
-import hydra
-
-from ..base.constants import MODEL_MEMVAE
-from ..base.base_model import BaseModelVAE
-from ..base.distributions import Normal
-from ..base.representations import ProductOfExperts, MeanRepresentation
-
-class me_mVAE(BaseModelVAE):
-    r"""
-    Multimodal Variational Autoencoder (MVAE).
-
-    Loss optimises the ELBO term from the joint posterior distribution, as well as the separate ELBO terms for each view.
-    me_mVAE stands for multi ELBO Multimodal VAE
-
-    Args:
-    cfg (str): Path to configuration file. Model specific parameters in addition to default parameters:
-        model.beta (int, float): KL divergence weighting term.
-        model.join_type (str): Method of combining encoding distributions.
-        model.sparse (bool): Whether to enforce sparsity of the encoding distribution.
-        model.threshold (float): Dropout threshold applied to the latent dimensions. Default is 0.
-        encoder.default._target_ (multiae.architectures.mlp.VariationalEncoder): Type of encoder class to use.
-        encoder.default.enc_dist._target_ (multiae.base.distributions.Normal, multiae.base.distributions.MultivariateNormal): Encoding distribution.
-        decoder.default._target_ (multiae.architectures.mlp.VariationalDecoder): Type of decoder class to use.
-        decoder.default.init_logvar (int, float): Initial value for log variance of decoder.
-        decoder.default.dec_dist._target_ (multiae.base.distributions.Normal, multiae.base.distributions.MultivariateNormal): Decoding distribution.
-
-    input_dim (list): Dimensionality of the input data.
-    z_dim (int): Number of latent dimensions.
-
-    References
-    ----------
-    Wu, M., & Goodman, N.D. (2018). Multimodal Generative Models for Scalable Weakly-Supervised Learning. NeurIPS.
-
-    """
-
-    def __init__(
-        self,
-        cfg = None,
-        input_dim = None,
-        z_dim = None
-    ):
-        super().__init__(model_name=MODEL_MEMVAE,
-                        cfg=cfg,
-                        input_dim=input_dim,
-                        z_dim=z_dim)
-
-        self.join_type = self.cfg.model.join_type
-        if self.join_type == "PoE":
-            self.join_z = ProductOfExperts()
-        elif self.join_type == "Mean":
-            self.join_z = MeanRepresentation()
-
-    def encode(self, x):
-        r"""Forward pass through encoder networks.
-
-        Args:
-            x (list): list of input data of type torch.Tensor.
-
-        Returns:
-            (list): Single element list of joint encoding distribution.
-        """
-        mu = []
-        logvar = []
-        for i in range(self.n_views):
-            mu_, logvar_ = self.encoders[i](x[i])
-            mu.append(mu_)
-            logvar.append(logvar_)
-        mu = torch.stack(mu)
-        logvar = torch.stack(logvar)
-        mu_out, logvar_out = self.join_z(mu, logvar)
-        qz_x = hydra.utils.instantiate( 
-            self.cfg.encoder.default.enc_dist, loc=mu_out, scale=logvar_out.exp().pow(0.5)
-        )
-        return [qz_x]
-
-    def encode_separate(self, x):
-        r"""Forward pass through encoder networks.
-
-        Args:
-            x (list): list of input data of type torch.Tensor.
-
-        Returns:
-            (list): list of separate encoding distributions for each modality.
-        """
-        qz_xs = []
-        for i in range(self.n_views):
-            mu_, logvar_ = self.encoders[i](x[i])
-            qz_x = hydra.utils.instantiate( 
[…]

[…]
