### 2023-04-11 63f18b7
update fit function to allow data list input and remove saved data from data loader
