### 2023-04-14 #5093 @arfon
@abhi-glitchhg, @Saran-nns – thanks for your reviews here. Could you please help with with a couple of things:

1. Provide links here to any issues you opened on the upstream repository (https://github.com/alawryaguila/multi-view-AE)
2. @abhi-glitchhg – it looks like you've completed your checklist. Could you also give me an indication here whether you're recommending we accept this submission?

Many thanks!
