### 2023-04-14 16972c4
fix model loading from ckpt and predict function with different datamodules
