+++ docs/JOSS/paper.md
@@ -15,6 +15,7 @@ authors:
   - name: Alejandra Jayme
     affiliation: 2
   - name: Nina Montaña-Brown
+  - orchid: 0000-0001-5685-971X
     affiliation: 1
   - name: Vincent Heuveline
     affiliation: 2
