+++ docs/JOSS/paper.md
@@ -18,6 +18,7 @@ authors:
   - orchid: 0000-0001-5685-971X
     affiliation: 1
   - name: Vincent Heuveline
+  - orchid: 0000-0002-2217-7558
     affiliation: 2
   - name: Andre Altmann
     orchid: 0000-0002-9265-2393
