### 2023-04-27 #5093 @alawryaguila
@arfon that's great!

I have now added ORCIDs to the paper.md file for all authors which have them.
I have also created a release of the software on github which has been uploaded to zenodo with the DOI: 10.5281/zenodo.7871099

Please let me know if there's anything else I need to do.
