+++ multiviewae/base/constants.py
@@ -13,6 +13,7 @@ MODEL_MVTCAE    = "mvtCAE"
 MODEL_DVCCA     = "DVCCA"
 MODEL_MOPOEVAE  = "MoPoEVAE"
 MODEL_MMJSD     = "mmJSD"
+MODEL_WEIGHTEDMVAE = "weighted_mVAE"
 MODELS = [
             MODEL_AE,
             MODEL_AAE,
@@ -26,7 +27,8 @@ MODELS = [
             MODEL_MVTCAE,
             MODEL_DVCCA,
             MODEL_MOPOEVAE,
-            MODEL_MMJSD
+            MODEL_MMJSD,
+            MODEL_WEIGHTEDMVAE
         ]
 
 ADVERSARIAL_MODELS = [
@@ -44,7 +46,8 @@ VARIATIONAL_MODELS = [
             MODEL_MVTCAE,
             MODEL_DVCCA,
             MODEL_MOPOEVAE,
-            MODEL_MMJSD
+            MODEL_MMJSD,
+            MODEL_WEIGHTEDMVAE
         ]
 
 SPARSE_MODELS = [
