+++ multiviewae/models/__init__.py
@@ -11,6 +11,7 @@ from .me_mvae import *
 from .jmvae import *
 from .mopoevae import *
 from .mmJSD import *
-__all__ = ["AE", "AAE", "jointAAE", "wAAE", "mcVAE", "mVAE", "JMVAE", "me_mVAE", "mmVAE", "mvtCAE", "DVCCA", "MoPoEVAE", "mmJSD"]
+from .weighted_mvae import *
+__all__ = ["AE", "AAE", "jointAAE", "wAAE", "mcVAE", "mVAE", "JMVAE", "me_mVAE", "mmVAE", "mvtCAE", "DVCCA", "MoPoEVAE", "mmJSD", "weighted_mVAE"]
 classes = __all__ 
 #TODO: use constants instead
