+++ multiviewae/base/dataloaders.py
@@ -1,11 +1,8 @@
 import random
 import numpy as np
-import pandas as pd
 import pytorch_lightning as pl
-
-from torchvision import transforms
+import hydra
 from torch.utils.data import DataLoader
-from .datasets import MVDataset
 
 class MultiviewDataModule(pl.LightningDataModule):
     """LightningDataModule for multi-view data.
@@ -22,6 +19,7 @@ class MultiviewDataModule(pl.LightningDataModule):
             batch_size,
             is_validate,
             train_size,
+            dataset,
             data,
             labels
         ):
@@ -32,20 +30,19 @@ class MultiviewDataModule(pl.LightningDataModule):
         self.train_size = train_size
         self.data = data  
         self.labels = labels 
-
+        self.dataset = dataset
         if not isinstance(self.batch_size, int):
             self.batch_size = self.data[0].shape[0]
 
     def setup(self, stage):
         if self.is_validate:
             train_data, test_data, train_labels, test_labels = self.train_test_split()
-            self.train_dataset = MVDataset(data=train_data, labels=train_labels)
-            self.test_dataset = MVDataset(data=test_data, labels=test_labels)
+            self.train_dataset = hydra.utils.instantiate(self.dataset, data=train_data, labels=train_labels)
+            self.test_dataset = hydra.utils.instantiate(self.dataset, data=test_data, labels=test_labels)
         else:
-            self.train_dataset = MVDataset(data=self.data, labels=self.labels) 
+            self.train_dataset = hydra.utils.instantiate(self.dataset, data=self.data, labels=self.labels) 
             self.test_dataset = None
-        #delete data from memory
-        del self.data
+            del self.data
 
     def train_test_split(self):
 
@@ -66,6 +63,69 @@ class MultiviewDataModule(pl.LightningDataModule):
             test_labels = self.labels[test_idx]
 
         return train_data, test_data, train_labels, test_labels
+
+    def train_dataloader(self):
+        return DataLoader(self.train_dataset, batch_size=self.batch_size, shuffle=False, num_workers=0) #use default num_workers for now, problem in windows! 
+
+    def val_dataloader(self):
+        if self.is_validate:
+            return DataLoader(
+                self.test_dataset, batch_size=self.batch_size, shuffle=False, num_workers=0)
+        return None
+
+
+class ListDataModule(pl.LightningDataModule):
+    """LightningDataModule for multi-view data.
+
+    Args:
+        batch_size (int): Batch size.
+        is_validate (bool): Whether to use a validation set.
+        train_size (float): Proportion of batch to use for training between 0 and 1. Remainder of batch is used for validation.
+        data (list): Input data. list of identifiers to load data from.
+        labels (np.array): Dataset labels. 
+    """
+
+    def __init__(
+            self,
+            batch_size,
+            is_validate,
+            train_size,
+            dataset,
+            data,
+            labels,
+        ):
+
+        super().__init__()
+        self.batch_size = batch_size
+        self.is_validate = is_validate
+        self.train_size = train_size
+        self.data = data[0]
+        self.labels = labels
+        self.dataset = dataset
+        if not isinstance(self.batch_size, int):
+            self.batch_size = len(data[0])
+
+    def train_test_split(self):
+
+        N = len(self.data)
+        train_idx = list(random.sample(range(N), int(N * self.train_size)))
+        test_idx = list(set(list(range(N))) -  set(train_idx))
+        data = self.data
+
+        train_data = [data[i] for i in train_idx]
+        test_data = [data[i] for i in test_idx]
+
+        return train_data, test_data
+
+    def setup(self, stage):
+
+        if self.is_validate:
+            train_data, test_data, = self.train_test_split()
[…]
