### 2023-05-16 #5093 @alawryaguila
> @alawryaguila – apologies for the delayed response here. Looks like there are some errors with the metadata – can you please merge this PR to fix? [alawryaguila/multi-view-AE#35](https://github.com/alawryaguila/multi-view-AE/pull/35)

Hi @arfon I have now merged, hopefully that fixes the issue :)
